Assignments/Task-3/22671A7370/requirements.txt
streamlit>=1.24.1
opencv-python>=4.7.0
numpy>=1.25.0
Pillow>=10.0.0
matplotlib>=3.8.0
